Better Help Function
Help function now allows arbitrary input. If no verb is given, the
function returns a message with all verbs and their descriptions.

If a verb is given and it is a supported verb, that verb and its
description is returned. If its not supported, all supported verbs with
the same first character are returned.