Start to Create Items with Adjectives
New functions that create objects in item classes using the respective
adjectives. Start sword and shield are currently the only items.

Issue: Player can take multiple of the same item. May need to
incorporate rooms and items in rooms to mitigate this issue. Or modify
the take command to only take new items.