Room Sweep and Room Dictionary
Created a sweep dictionary where the key is a value between 1 and
roomsX*roomsY corresponding to the position (x,y) where sweep[i] =
x+y*roomsX

Created a dictionary also called rooms with keys as the first value and
class RoomClass as the value. Should allow me to place items and enemies
in a specific room with
```
RM.rooms[i].items
```
Maybe.