Move: First Draft
MAJOR CHANGE: Player name now stored as 'you' in class dictionaries.
This allows an easy access key to make any changes to the player. A
catch was added in the getInput function to alter the incoming noun if
it was the player's designated name and make the switch to 'you'. This
could cause some issues to pull and merge downstream, but long term, I
think this is a good choice.

Introducing the move command! Players can enter "move" to see possible
compass directions and then move in those directions.

Player location also added to description.