Base Adjectives and Modified Hit
Enemies now have base value for health, attack, and defense. These are
declared upon instantiation of the character, so there may be some
problems modifying them for harder enemies. But, if we can modify them
with item attributes, as the goal of the super to this branch, then the
same method could be applied to increasing attributes of enemies by
location.

Hit routine also uses player attack and enemy defense to compute the
damage dealt to the enemy.