Difficulty Map and Spawning Scheme
Hardcoded a difficulty map, player spawn, and boss location into the
game. Game WILL Create a 5x5 game space with the player in one corner
(1,1) and the boss in the opposite corner (5,5). Rooms will get more
difficult and rewarding as the player approaches the boss according to
the list of lists roomDiff.

PDF contains flowchart of the scheme for randomly populating rooms with
items and enemies of various types and sizes according to room
difficulty. Will need to include a description of determining what
items/enemy types to spawn.