Better Item Creation
Fixes #15 by only creating items that have adjectives left to create. If
the number of adjectives in an item type is greater than zero, the item
class and adjectives are added to a dictionary. It is from this
dictionary that the item type and name are selected. Therefore, if there
are no more adjectives remaining in an item type, another item type will
be selected. No more item thresholds. A similar technique will be
applied to the enemies method. This will allow the removal of the
rareThresh dictionary.

Also removed calls to RM.addToRoom in the __init__ method for items.