:hocho: Tic-for-tat combat
**Better combat mechanics through object-oriented programming!**

If the player successfully attacks an enemy, the enemy attempts to hit
back. Successful attacks happen when the attacker has a higher attack
than the target's defense. If this is not the case, a random number is
drawn. If the random number is greater than 0.5, the attack is
successful.

Also, there is a placeholder for if the player's health drops below
zero.