Better Input Handling
Each verb function now takes in two arguments, with the second being defaulted to None. This allows easier retrofitting of the "[verb] trusty sword" inputs without breaking other functions that only require one argument. 

There is some try/except catching that goes on in the get input that streamlined a lot of the code too.