Two Argument Take
Take now uses two arguments, item adjective and item type, to add a specified item to the player's pack. Need to update the equip features next. The player's pack now uses the adjective and the item type as the keys to a given item object value.

Todo - Items should be added to rooms with keys as the item adjective and item type. i.e. RM.rooms[swp].items["trusty sword"]  = ItemMod.Sword object blah blah blah. This will allow multiple items with the same adjective to be added to rooms.