Unique Room Items and Better Take
Items are added to the room with their adjective and item type as the keys to their values.  This allows items of different types with the same adjective to be added to the same room, and for the player to grab these unique items.

Todo - Work on the equip/unequip routines. They need to handle two argument inputs.