MAJOR: Restart Upon Death and Better Game Creation
Room population and adjective building now take place in sepearate funcitons, called at the beginning of the game. This allows the game to be restarted and any deleated adjectives to be recreated at the start of any game. 
If the player loses all their health in a fight, a game over screen is launched. The player is given the option to restart the game or not. If the game is restarted, all rooms are cleared, and then all adjective and room dictionaries recreated. The player is given the same start of game prompts, minus the giant TBPG and help commands.