MAJOR: Better Player Method and Objects
All references to "CM.GameCharacter.objects[you]" to obtain the player's class and thus attributes have been replaced with "VFM.player" or "palyer" inside of verbFuncMod. The object of the player is created in the room sweep still, and stored as a variable inside of VFM, as it is refered to there the most. 

Also removed the getLoc() function. The players sweep location is obtained via "VFM.player.sweep" now. Waaaaay simpler.

Updated functions:

- equip
- take
- hit
- unequip
- drink
- examine